del empty folder, progress bar, test data

+++ Photo_sort_exif.py
@@ -1,10 +1,15 @@
+from itertools import count
 import os
-import re
+import shutil
+from sre_parse import State
 from exif import Image
 from os.path import getmtime
 from datetime import datetime as dt
-import shutil
-from os import path
+from progress.bar import Bar
+
+
+
+
 
 def create_move(path, image):
     os.makedirs(path,exist_ok=True)
@@ -19,7 +24,7 @@ def create_move(path, image):
         elif os.stat(image).st_size != os.stat(path + "\\" + image.split("\\")[-1]).st_size:
             new_path = path + "\\" + (image.split("\\")[-1]).split('.')[-2] + '_offsize_.' + (image.split("\\")[-1]).split('.')[-1]
             shutil.move(image, new_path )
-            print(f'{image} сохранен как {new_path}')
+            print(f'{image} сохранен как {new_path}')            
             
             
 def remove_empty_dir(path_input:str, del_mode:int ):
@@ -28,14 +33,16 @@ def remove_empty_dir(path_input:str, del_mode:int ):
         for i in os.walk(path_input):
             all_dir.insert(0,i)
         for dir in all_dir:
-            if dir[-1] == []:
+            if dir[-1] == [] and dir[-2] == []:
                 os.removedirs(dir[0])
             else:
-                print(f'В папке {dir} имеются файлы отличные от изображений')
+                #print(f'В папке {dir} имеются файлы отличные от изображений')
                 continue
     elif del_mode != 0:
         print(f'Выбран неверный режим удаления {del_mode}, '
-              f'Пакпи небыли удалены') 
+              f'Пакпи небыли удалены')
+    elif del_mode == 0:
+        print(f'Выбран режим без удаления ')
           
 
 def create_path_exif(img_date, path_out, mode):
@@ -72,7 +79,7 @@ def create_path_win(image, path_out, mode):
         print(f"Введён неверный режим работы, {mode} - недопустимое значение")
 
 
-def sort_photo(path_input, path_out, mode_parse): #
+def sort_photo(path_input, path_out, mode_parse, del_mode):
     # указываем расширения файлов для поиска и обработки
     type_file = ('jpg','JPG','NEF', 'nef', 'RAW', 'raw', 'PNG', 'png')
     # в цикле собираем все файлы и дириктории
@@ -90,16 +97,16 @@ def sort_photo(path_input, path_out, mode_parse): #
                 # # Если информация о дате съемке присутсвует то обрабатываем
                         if img_date.get('datetime')!= None:
                             create_move(create_path_exif(img_date, path_out, mode_parse), image)
+                    else:
+                        create_move(create_path_win(image, path_out, mode_parse), image)
+                        
                 except Exception as e:
                     print(f'{e} хз что не так {image}')
-                finally:
-                    create_move(create_path_win(image, path_out, mode_parse), image)
-    
+                    create_move(create_path_win(image, path_out, mode_parse), image)       
 
-                     
-                                                   
-
-
+    remove_empty_dir(path_input, del_mode)
+ 
+   
 def main():
     mode_parse = int(input(f'выбор режима раскладывания по каталогам \n'
                 f'1- все в кучу (в папку path_out)\n'
@@ -108,15 +115,23 @@ def main():
                 f'4- в папке с годами создает папку с месяцем и в ней разбивка по дням\n'
                 f'Введите режим обработки: '))
     del_mode = int(input(f"Режим удаления пустых папок после перемещния фото \n"
+                        f'0- папки не удаляются\n'
                         f'1- удаляет все пустые папки\n'
-                        f'0- папки не удаляются \n'
                         f'Введите режим удаления: '))
     # указываем папку для поиска и обработки фото
     path_input = input("Введите путь для обрабатываемой папки: ")      
 
     # указываем папку для сохранения результат
-    path_out = input("Введите путь для сохранения фото: ") 
-    sort_photo(path_input, path_out, mode_parse) 
+    path_out = input("Введите путь для сохранения фото: ")
+
+    while sort_photo(path_input, path_out, mode_parse, del_mode):
+        with Bar('Processing') as bar:
+            for i in range(20):
+        # Do some work
+                bar.next()
+
+    
+    print('Все готово!')
